server/requirements.txt
Flask==3.0.3
pymongo
flask-cors
python-dotenv
Werkzeug
